refactor: extract sandboxed Jinja templating into widgets/_template
Move the HTML widget's template context, renderer, and entity-reference
scanning into a shared module so other widget fields (labels, text
content) can reuse the exact same sandboxed Jinja subset. Also make
`now` callable (`now()`) so templates pasted from HA automations work
unchanged alongside the documented `now` variable.
