+++ custom_components/geekmagic/widgets/html.py
@@ -10,13 +10,11 @@ theme chrome, so user templates can use ``var(--text-primary)``,
 from __future__ import annotations
 
 import logging
-import re
 from html import escape
 from typing import TYPE_CHECKING, Any, ClassVar
 
-from jinja2.sandbox import SandboxedEnvironment
-
 from ..htmldoc import mdi_span
+from ._template import _render_template, template_entity_refs
 from .base import Widget, WidgetConfig
 
 if TYPE_CHECKING:
@@ -24,61 +22,6 @@ if TYPE_CHECKING:
     from .state import WidgetState
 
 _LOGGER = logging.getLogger(__name__)
-
-# Entity references inside the Jinja template, e.g. states('sensor.temp'),
-# state_attr("climate.living", "current_temperature") or
-# is_state('light.kitchen', 'on'). Used to declare entity dependencies so
-# the coordinator pre-fetches them into WidgetState.
-_ENTITY_REF_RE = re.compile(r"""\b(?:states|state_attr|is_state)\(\s*['"]([^'"]+)['"]""")
-
-
-def _build_template_context(
-    state: WidgetState, primary_entity_id: str | None = None
-) -> dict[str, Any]:
-    """Build the Jinja context exposed to the user's HTML template.
-
-    The convenience variables (``state``/``name``/``unit``/...) come from
-    ``primary_entity_id`` when given (the widget's "Entity (template
-    data)" selector, delivered by the coordinator in
-    ``WidgetState.entities``), falling back to ``state.entity``.
-    """
-
-    def states(entity_id: str) -> str:
-        entity = state.get_entity(entity_id)
-        return entity.state if entity else "unknown"
-
-    def state_attr(entity_id: str, attribute: str) -> Any:
-        entity = state.get_entity(entity_id)
-        return entity.get(attribute) if entity else None
-
-    def is_state(entity_id: str, value: str) -> bool:
-        return states(entity_id) == value
-
-    entity = state.get_entity(primary_entity_id) if primary_entity_id else None
-    if entity is None:
-        entity = state.entity
-    return {
-        "entity": entity,
-        "state": entity.state if entity else "",
-        "name": entity.friendly_name if entity else "",
-        "unit": entity.unit if entity else "",
-        "attributes": entity.attributes if entity else {},
-        "now": state.now,
-        "states": states,
-        "state_attr": state_attr,
-        "is_state": is_state,
-    }
-
-
-def _render_template(source: str, state: WidgetState, primary_entity_id: str | None = None) -> str:
-    """Render the user's Jinja template against widget state.
-
-    Uses a sandboxed environment: templates come from the user's own HA
-    config, but sandboxing is cheap insurance against accidental access
-    to Python internals.
-    """
-    env = SandboxedEnvironment(autoescape=False)
-    return env.from_string(source).render(**_build_template_context(state, primary_entity_id))
 
 
 def _placeholder(message: str, icon: str = "code-tags") -> str:
@@ -193,7 +136,7 @@ class HtmlWidget(Widget):
             entities.append(self.config.entity_id)
         if self.dynamic_entity_id and self.dynamic_entity_id not in entities:
             entities.append(self.dynamic_entity_id)
-        for entity_id in _ENTITY_REF_RE.findall(self.html):
+        for entity_id in template_entity_refs(self.html):
             if entity_id not in entities:
                 entities.append(entity_id)
         return entities
